[FIX] moved patch_and_pad_image to module utils.predict_patches
